==> R/00_doit.R <==
# Run all scripts --------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_model_generel.R")
source(file = "R/05_model_HM.R")
source(file = "R/06_model_PCA.R")
source(file = "R/07_model_pvalue.R")
rmarkdown::render(input = "/cloud/project/doc/Presentation.Rmd")
